requirements.txt
flask>=3.0
azure-identity
azure-ai-projects>=1.0.0
openai
requests
python-dotenv
